add test claim credits page
